Refactor : types 폴더 이동
