+++ src/pages/playlist/Modify.tsx
@@ -16,7 +16,7 @@ import { toast } from "react-toastify";
 import { useYoutubeInfo } from "@/pages/playlist/hooks/useYoutubeInfo";
 import { useThumbnail } from "@/pages/playlist/hooks/useThumbnailUpload";
 import { getPlaylistWithVideos } from "@/shared/api/playlistWithvideos";
-import { Video } from "@/types/video";
+import { Video } from "@/shared/types/video";
 import { useUpdatePlaylist } from "@/pages/playlist/hooks/useUpdatePlaylist";
 import ErrorFallback from "@/shared/component/ErrorFallback";
